// src/imports/enums/states.ts
export enum ESettingsSections {
    APPLICATION = 'application',
    CONTENT = 'content',
}

export enum ECardReducerStates {
    LEVEL = 'cardLevel',
    GAME_TIME = 'cardGameTime',
    BACKGROUND_DIR = 'cardBackgroundDir',
    MENU_OPEN = 'cardMenuOpen',
    MENU_ORIGIN = 'cardMenuOrigin',
    CARD_PRESSED = 'cardPressed',
}

export enum ESettingsStates {
    LIST_LAYOUT = 'listLayout',
    AUTO_REFRESH = 'autoRefresh',
    REFRESH_INTERVAL = 'refreshInterval',
    SHOW_SUMMONER_IDS = 'showSummonerIds',
    SHOW_RANDOM_SKINS = 'showRandomSkins',
    SHOW_TEAM_LOGOS = 'showTeamLogos',
    SHOW_UNAVAILABLE = 'showUnavailable',
    APP_THEME = 'appTheme',
    APP_SCALE = 'appScale',
    OPEN_ON_STARTUP = 'openOnStartup',
    MINIMIZE_TO_TRAY = 'minimizeToTray',
    HARDWARE_ACCELERATION = 'hardwareAcceleration',
    ANIMATIONS = 'showAnimations',
    USE_BACKGROUND = 'useBackground',
    RANDOM_BACKGROUND = 'randomBackground',
    LIVE_BACKGROUND = 'liveBackground',
    KEYBOARD_MODE = 'keyboardMode',
    NOTIFICATIONS = 'notifications',
    APP_LANGUAGE = 'appLanguage',
}